+++ tools/eslint-config/rules/prettier.js
@@ -1,6 +1,7 @@
 module.exports = {
   extends: ['plugin:prettier/recommended', 'prettier/@typescript-eslint'],
   rules: {
+    curly: ['error', 'all'],
     'prettier/prettier': [
       'error',
       {